requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
discord.py==2.3.2
frozenlist==1.4.0
idna==3.6
multidict==6.0.4
mysql-connector-python==8.2.0
protobuf==4.21.12
python-dotenv==1.0.0
yarl==1.9.4
